add getAllBalance, fee discount for nft holders, update script and refactor

+++ contracts/interfaces/IUniswapV2Factory.sol
@@ -12,7 +12,7 @@ interface IUniswapV2Factory {
 
     function getPair(address tokenA, address tokenB) external view returns (address pair);
     function allPairs(uint) external view returns (address pair);
-    function allPairsLength() external view returns (uint);
+    function allPairsLength() external view returns (uint256);
 
     function createPair(address tokenA, address tokenB) external payable returns (address pair);
